Refactor eraser
